pyproject.toml
[project]
name = "multimodal-iad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anomalib[core, vlm, openvino]>=2.0.0",
    "hydra-zen>=0.15.0",
    "pydantic>=2.11.7",
    "PyQt6>=6.9.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "strenum>=0.4.15",
    "timm==1.0.10",
    "scikit-learn>=1.7.0",
    "openvino>=2025.2",
    "python-dotenv>=1.1.1",
]

[project.scripts]
multimodal-iad = "multimodal_iad.gui.app:main"

[tool.uv]
package = true

[dependency-groups]
dev = ["ruff>=0.12.0"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
fixable = ["ALL"]
ignore = ["D203", "D213", "COM812"]

[tool.ruff.format]
docstring-code-format = true
